Many changes in the lexer design
- Reverted the decision to delete the null token.
- Used the standard <ctype.h> instead of some equivalent functions
  defined in "lexer_utils.c" (NB: `isXdigit` follows the naming pattern
of the standard functions because hexadecimal is mainstream enough to not
be considered as a special case of the compiler, obviously).
- Finally supports the multiple qualifiers scenario.
- Simplified lexer errors checking by using `get_next_word` instead
  of iterating through each single character of the code within the process.
- Now, specific errors must be referenced using a label. It would be
  easier to comprehend where they could occur in the future (see
"lexer_error.c" new comments for illustration).

The design of the lexer has been changed. It no longer designates what
are keys, lock and keywords but instead look at what is around
a colon and marks it as L (Left) or R (Right). This will be more
consistent with the parsing, which will manage this.