Major commit for parameterized label identification support
- Better parameterized label debugging.
- Increased the limit up to eight built-in qualifiers (right and left)
  instead of four of each type. This limit will be removed later.
- Simplified the macros to define enumeration for more clarity.
- The node type `NodeType_CHILD` was removed because the first node of
  an instruction is probably enough to determine the behaviour of the
children. Moreover, it adds the possibility for children to have a subtype.
- Support key types for the children of identification nodes. It allows to
  get to know if we process a modifier or a lock. The subtype make it
possible to identify lonely locks and locks of parameterized labels.
- Lexers now produce R left parenthesis and R right parenthesis. It
  makes it easier to parse parameterized labels.

Before this commit, only parameterized labels returning a lock and taking
at least one parameter were supported. The new design of the lexer and
the identification part of the parser has been slightly changed.